consolidate the YouTubePlayerView binding adapters
