+++ ui/base/src/main/java/org/cru/godtools/base/ui/databinding/YouTubePlayerViewBindingAdapter.kt
@@ -5,7 +5,9 @@ import androidx.databinding.adapters.ListenerUtil
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.PlayerConstants
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.YouTubePlayer
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.listeners.AbstractYouTubePlayerListener
+import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.listeners.YouTubePlayerCallback
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.listeners.YouTubePlayerListener
+import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.utils.YouTubePlayerTracker
 import com.pierfrancescosoffritti.androidyoutubeplayer.core.player.views.YouTubePlayerView
 import org.cru.godtools.base.ui.R
 
@@ -26,3 +28,27 @@ fun YouTubePlayerView.recueVideo(recue: Boolean) {
         current != null -> removeYouTubePlayerListener(current)
     }
 }
+
+@BindingAdapter("videoId")
+fun YouTubePlayerView.updateVideo(videoId: String?) = updateVideo(videoId, false)
+
+@BindingAdapter("videoId", "autoPlay")
+fun YouTubePlayerView.updateVideo(videoId: String?, autoPlay: Boolean) {
+    getYouTubePlayerWhenReady(object : YouTubePlayerCallback {
+        override fun onYouTubePlayer(youTubePlayer: YouTubePlayer) {
+            when {
+                tracker.videoId == videoId -> Unit
+                videoId == null -> youTubePlayer.pause()
+                autoPlay -> youTubePlayer.loadVideo(videoId, 0f)
+                else -> youTubePlayer.cueVideo(videoId, 0f)
+            }
+        }
+    })
+}
+
+private val YouTubePlayerView.tracker
+    get() = ListenerUtil.getListener<YouTubePlayerTracker>(this, R.id.youtubeplayer_tracker)
+        ?: YouTubePlayerTracker().also {
+            ListenerUtil.trackListener(this, it, R.id.youtubeplayer_tracker)
+            addYouTubePlayerListener(it)
+        }
