+++ ui/base/src/main/res/values/ids.xml
@@ -5,5 +5,6 @@
 
     <!-- Data Binding Listeners -->
     <item name="youtubeplayer_recue_listener" type="id" />
+    <item name="youtubeplayer_tracker" type="id" />
     <item name="lifecycleOwner" type="id" />
 </resources>
